deniyoruz/requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
mysqlclient==2.2.0
pandas==2.0.3
openpyxl==3.1.2
gunicorn
python-dotenv
